+++ roadeye-api/roadeye-api-hub/src/main/java/org/re/mdtlog/service/MdtCycleLogService.java
@@ -6,19 +6,19 @@ import org.re.common.api.payload.MdtLogRequestTimeInfo;
 import org.re.mdtlog.domain.TransactionUUID;
 import org.re.mdtlog.dto.MdtCycleLogMessage;
 import org.re.mdtlog.dto.MdtEventMessage;
-import org.re.messaging.MessagingService;
 import org.re.messaging.amqp.AMQPQueue;
+import org.re.messaging.amqp.AMQPService;
 import org.springframework.stereotype.Service;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class MdtCycleLogService {
-    private final MessagingService messagingService;
+    private final AMQPService AMQPService;
 
     public void addCycleLogs(TransactionUUID tuid, MdtCycleLogMessage dto, MdtLogRequestTimeInfo timeInfo) {
         var queueName = AMQPQueue.QueueNames.MDT_CAR_LOCATION;
         var message = new MdtEventMessage<>(tuid.toString(), dto, timeInfo.sentAt(), timeInfo.receivedAt());
-        messagingService.sendWithQueueName(queueName, message);
+        AMQPService.sendWithQueueName(queueName, message);
     }
 }
