+++ roadeye-api/roadeye-api-hub/src/main/java/org/re/mdtlog/service/MdtIgnitionService.java
@@ -7,25 +7,25 @@ import org.re.mdtlog.domain.TransactionUUID;
 import org.re.mdtlog.dto.MdtEventMessage;
 import org.re.mdtlog.dto.MdtIgnitionOffMessage;
 import org.re.mdtlog.dto.MdtIgnitionOnMessage;
-import org.re.messaging.MessagingService;
 import org.re.messaging.amqp.AMQPQueue;
+import org.re.messaging.amqp.AMQPService;
 import org.springframework.stereotype.Service;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class MdtIgnitionService {
-    private final MessagingService messagingService;
+    private final AMQPService AMQPService;
 
     public void ignitionOn(TransactionUUID tuid, MdtIgnitionOnMessage dto, MdtLogRequestTimeInfo timeInfo) {
         var queueName = AMQPQueue.QueueNames.MDT_IGNITION_ON;
         var message = new MdtEventMessage<>(tuid.toString(), dto, timeInfo.sentAt(), timeInfo.receivedAt());
-        messagingService.sendWithQueueName(queueName, message);
+        AMQPService.sendWithQueueName(queueName, message);
     }
 
     public void ignitionOff(TransactionUUID tuid, MdtIgnitionOffMessage dto, MdtLogRequestTimeInfo timeInfo) {
         var queueName = AMQPQueue.QueueNames.MDT_IGNITION_OFF;
         var message = new MdtEventMessage<>(tuid.toString(), dto, timeInfo.sentAt(), timeInfo.receivedAt());
-        messagingService.sendWithQueueName(queueName, message);
+        AMQPService.sendWithQueueName(queueName, message);
     }
 }
