+++ roadeye-api/roadeye-api-hub/src/main/java/org/re/messaging/amqp/AMQPService.java
@@ -1,4 +1,4 @@
-package org.re.messaging;
+package org.re.messaging.amqp;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class MessagingService {
+public class AMQPService {
     private final RabbitTemplate rabbitTemplate;
 
     public void sendWithQueueName(String queueName, Object message) {
